backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
pytest
pytest-asyncio
respx